requirements.txt
alembic==1.14.0
antiorm==1.2.1
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
db==0.1.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.6
pycparser==2.22
PyJWT==2.3.0
python-dotenv==1.0.1
qrcode==8.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
